Split token refresh and retry policy into different middlewares. Corrected MethodTemplate.
