+++ AutoRest/Generators/Ruby/Ruby/Templates/MethodTemplate.cshtml
@@ -1,4 +1,5 @@
-﻿@using System.Linq;
+﻿@using System
+@using System.Linq;
 @using Microsoft.Rest.Generator.ClientModel
 @using Microsoft.Rest.Generator.Ruby.TemplateModels
 @inherits Microsoft.Rest.Generator.Template<Microsoft.Rest.Generator.Ruby.MethodTemplateModel>
@@ -28,8 +29,11 @@ def @(Model.Name)(@(Model.MethodParameterDeclaration))
   @(Model.RemoveDuplicateForwardSlashes("url"))
 
   @EmptyLine
+  @* TODO: add proper references to @client *@
   connection = Faraday.new(:url => url) do |faraday|
     faraday.use RetryPolicyMiddleware, times: 3, retry: 0.02, credentials: @(Model.ClientReference).credentials
+    faraday.use TokenRefreshMiddleware, credentials: @(Model.ClientReference).credentials
+    faraday.use :cookie_jar
     faraday.adapter Faraday.default_adapter
   end
 
@@ -68,7 +72,10 @@ def @(Model.Name)(@(Model.MethodParameterDeclaration))
   promise = Concurrent::Promise.new do
     connection.put do |request|
       request.headers = request_headers
-      request.body = request_content
+@if (Model.RequestBody != null)
+{
+      @:request.body = request_content
+}
     end
   end
 
@@ -80,17 +87,18 @@ def @(Model.Name)(@(Model.MethodParameterDeclaration))
       @if (Model.DefaultResponse != null)
       {
       @:error_model = JSON.load(response_content)
-      @:fail @(Model.OperationExceptionTypeString).new(request_content, http_response, error_model)
+      @:fail @(Model.OperationExceptionTypeString).new(http_response, http_response, error_model)
       }
       else
       {
-      @:fail @(Model.OperationExceptionTypeString).new(request_content, http_response)
+      @:fail @(Model.OperationExceptionTypeString).new(http_response, http_response)
       }
     end
 
     @EmptyLine
+    @* TODO: add propert request object into HttpOperationResponse and HttpOperationException *@
     # Create Result
-    result = @(Model.OperationResponseReturnTypeString).new(request_content, http_response)
+    result = @(Model.OperationResponseReturnTypeString).new(http_response, http_response)
     @Model.InitializeResponseBody
 
     @foreach (var responsePair in Model.Responses.Where(r => r.Value != null && r.Value.IsSerializable()))
