+++ ClientRuntimes/Ruby/ms-rest-azure/lib/ms_rest_azure/polling_state.rb
@@ -34,10 +34,10 @@ module MsRestAzure
   	  if (!@resource.nil? && @resource.respond_to?(:properties) && @resource.properties.respond_to?(:provisioning_state) && !@resource.properties.provisioning_state.nil?)
   	  	@status = @resource.properties.provisioning_state
   	  else
-  	  	case @response.code
-  	  	  when "202"
+  	  	case @response.status
+  	  	  when 202
   	  	    @status = AsyncOperationStatus::IN_PROGRESS_STATUS
-  	  	  when "200", "201", "204"
+  	  	  when 200, 201, 204
   	  	    @status = AsyncOperationStatus::SUCCESS_STATUS
   	  	  else
   	  	  	@status = AsyncOperationStatus::FAILED_STATUS
