+++ ClientRuntimes/Ruby/ms-rest/lib/ms_rest.rb
@@ -15,6 +15,7 @@ require 'ms_rest/serialization.rb'
 require 'ms_rest/http_operation_response'
 require 'ms_rest/http_operation_exception'
 require 'ms_rest/retry_policy_middleware'
+require 'ms_rest/token_refresh_middleware'
 require 'ms_rest/service_client'
 
 module MsRest; end
